+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+
+- Prevent repeated access of `Transform.position` in `Transform.Set` functions.
+
 ## [1.0.1] - 2024-07-10
 
 ### Fixed